+++ internal/server/hub.go
@@ -1783,6 +1783,8 @@ func (h *hub) run() {
 				h.carvePumpkin(players, e)
 			case evLightOre:
 				h.lightOre(players, e)
+			case evDragonEgg:
+				h.onDragonEgg(players, e)
 			case evUseWoodShelf:
 				h.useWoodShelf(players, e)
 			case evUseCandle: